posix_file_impl: Propagate and keep open_flags
Keep open_flags for posix files (and its children), as well as
posix file handle.
Note that we get the actual flags from the fd, not the ones
given to the open call. Just to be sure.
Also note that if any code takes upon itself to do a "reopen"
or similar, this will break.